crypto: inside-secure - Fix -EINVAL error for authsize=0 for AEAD ciphers
AEAD ciphers were returning -EINVAL for authsize=0 cases, which caused
testmgr to occasionally fail the driver on the fuzz testing.

While an ICV size of 0 does not seem to make a whole lot of sense (why
do authenticated encryption and throw away the authentication result?)
we do not want to be failed (and hence disabled) by testmgr.