crypto: inside-secure: Fix stability issue with Macchiatobin board(URGENT)
This patch fixes a serious stability issue witnessed on the Macchiatobin
board, resulting in the driver reporting a continuous stream of result
descriptor errors, caused by a sticky ECC error in the Transform Record
Cache due to reading past the fetched data. (This is a HW bug, the patch
ensures the record fetched is large enough to prevent ever exceeding what
was actually fetched at the expense of fetching a bit more than needed).